# Neo4j optimizations, part 1