This PR is part 1 of a two (or three) part series of query optimizations to CoGEx and takes care of switching to parameterized querying (as mentioned in issue #127).

The changes in this PR are centered around using parameters instead of literal values in queries.

Query with a literal value:
```cypher
MATCH (n:Person {name: 'Anna'})
RETURN n
```

The parameterized version of the same query:
```cypher
MATCH (n:Person {name: $name})
RETURN n
```
Instead of adding in literals to the query, the parameter `name` is passed to the transaction function (`do_cypher_tx` in CoGEx).

## Current speedup

~_Some tests are running slower after the update and would need another round of edits or simply be reverted to the older implementation using literal values_~

_**Edit 2023-01-18 11:54:00 PST**
Fixed name duplication in commit https://github.com/bgyori/indra_cogex/pull/129/commits/9f06dae96588e84ef4cce865dfe0687208d43009 and updated the table with new times_

|                                        |   main [s] |   branch [s] |   diff: branch-main [s] |   fraction speedup |
|:---------------------------------------|-------:|---------:|--------------:|-----------------:|
| test_get_stmts_for_pmid                |  17.48 |    16.18 |         -1.3  |           -0.074 |
| test_get_mesh_ids_for_pmid             |  17.3  |    16.15 |         -1.15 |           -0.066 |
| test_get_stmts_for_mesh_id_w_children  |   4.72 |     4.61 |         -0.11 |           -0.023 |
| test_drugs_for_target                  |   4.72 |     4.91 |          0.19 |            0.04  |
| test_get_pmids_for_mesh                |   4.72 |     4.65 |         -0.07 |           -0.015 |
| test_get_genes_in_tissue               |   4.57 |     3.52 |         -1.05 |           -0.23  |
| test_is_drug_target                    |   4.49 |     4.44 |         -0.05 |           -0.011 |
| test_get_evidence_obj_for_mesh_id      |   4.49 |     4.37 |         -0.12 |           -0.027 |
| test_targets_for_drug                  |   2.7  |     2.74 |          0.04 |            0.015 |
| test_get_stmts_for_mesh_id_wo_children |   2.57 |     2.49 |         -0.08 |           -0.031 |
| test_is_gene_in_tissue                 |   2.35 |     2.37 |          0.02 |            0.009 |
| test_get_ontology_parent_terms         |   2.35 |     2.26 |         -0.09 |           -0.038 |
| test_isa_or_partof                     |   2.34 |     2.25 |         -0.09 |           -0.038 |
| test_get_tissues_for_gene              |   2.33 |     2.26 |         -0.07 |           -0.03  |
| test_get_ontology_child_terms          |   2.32 |     2.26 |         -0.06 |           -0.026 |
| test_get_mesh_child_terms_nonempty     |   2.32 |     2.27 |         -0.05 |           -0.022 |
| test_get_mesh_child_terms_empty        |   2.31 |     2.26 |         -0.05 |           -0.022 |
| test_get_trials_for_disease            |   1.24 |     1.24 |          0    |            0     |
| test_get_stmts_by_hashes               |   1.03 |     0.92 |         -0.11 |           -0.107 |
| test_get_trials_for_drug               |   0.98 |     1.01 |          0.03 |            0.031 |
| test_get_shared_pathways_for_gene      |   0.78 |     0.78 |          0    |            0     |
| test_get_pathways_for_gene             |   0.7  |     0.7  |          0    |            0     |
| test_get_genes_for_go_term             |   0.69 |     0.66 |         -0.03 |           -0.043 |
| test_is_gene_in_pathway                |   0.64 |     0.66 |          0.02 |            0.031 |
| test_get_genes_for_pathway             |   0.64 |     0.64 |          0    |            0     |
| test_get_evidence_obj_for_stmt_hashes  |   0.6  |     0.61 |          0.01 |            0.017 |
| test_get_evidence_obj_for_stmt_hash    |   0.55 |     0.55 |          0    |            0     |
| test_get_diseases_for_trial            |   0.55 |     0.53 |         -0.02 |           -0.036 |
| test_get_drugs_for_trial               |   0.55 |     0.55 |          0    |            0     |
| test_is_gene_mutated                   |   0.51 |     0.51 |          0    |            0     |
| test_is_go_term_for_gene               |   0.49 |     0.48 |         -0.01 |           -0.02  |
| test_get_side_effects_for_drug         |   0.47 |     0.46 |         -0.01 |           -0.021 |
| test_is_side_effect_for_drug           |   0.47 |     0.46 |         -0.01 |           -0.021 |
| test_get_go_terms_for_gene             |   0.47 |     0.46 |         -0.01 |           -0.021 |
| test_get_drugs_for_side_effect         |   0.47 |     0.47 |          0    |            0     |
| test_drugs_for_targets                 | nan    |     4.89 |        nan    |          nan     |
| test_targets_for_drugs                 | nan    |     3.09 |        nan    |          nan     |